[#549] reconfigure entity matcher dir layout; revise cmp_person
