+++ lib/tasks/cmp.rake
@@ -32,19 +32,19 @@ namespace :cmp do
 
       file_path = Rails.root.join('data', 'cmp_individuals_with_match_info.csv').to_s
 
+
       people = Cmp::Datasets.people.values.map do |cmp_person|
-        attrs = LsHash.new(littlesis_entity_id: '').merge!(cmp_person.attributes_with_matches)
+        attrs = LsHash.new(littlesis_entity_id: '').merge!(cmp_person.attributes)
 
-        attrs['littlesis_entity_id'] = 'new' if attrs['match1_name'].blank?
+        #attrs['littlesis_entity_id'] = 'new' if attrs['match1_name'].blank?
 
-        org_names = Cmp::Datasets
-                          .relationships
-                          .select { |r| r.fetch(:cmp_person_id, nil) == attrs.fetch(:cmpid, nil) }
-                          .map { |r| r.fetch(:cmp_org_id) }
-                          .map { |id| Cmp::Datasets.orgs.find { |o| o.fetch(:cmpid).to_s == id } }
-                          .compact
-                          .map { |cmp_org| cmp_org.fetch(:cmpname, '') }
-                          .join('| ')
+        associated = CmpEntity
+                       .where(cmp_id:Cmp::Datasets.relationships
+                                .select { |r| r.fetch(:cmp_person_id, 'REL_ID') == attrs.fetch(:cmpid, "ATTR_ID") }
+                                .map { |r| r.fetch(:cmp_org_id) })
+                       .map(&:entity)
+        
+        
 
         attrs.merge!('associated_corps' => org_names)
       end
